ref(utils): Remove actix utilities
