--- relay-server/src/utils/actix.rs
@@ -1,11 +1,0 @@
-use tokio::runtime::Runtime;
-
-/// Constructs a tokio [`Runtime`] configured for running [services](relay_system::Service).
-pub fn create_runtime(name: &str, threads: usize) -> Runtime {
-    tokio::runtime::Builder::new_multi_thread()
-        .thread_name(name)
-        .worker_threads(threads)
-        .enable_all()
-        .build()
-        .unwrap()
-}
