+++ relay-server/src/utils/mod.rs
@@ -1,4 +1,3 @@
-mod actix;
 mod api;
 mod buffer;
 mod dynamic_sampling;
@@ -19,7 +18,6 @@ mod native;
 #[cfg(feature = "processing")]
 mod unreal;
 
-pub use self::actix::*;
 pub use self::api::*;
 pub use self::buffer::*;
 pub use self::dynamic_sampling::*;
